Removing explicit casts when logging enums

+++ src/engine/controller/audio_routing_controller.cpp
@@ -97,7 +97,7 @@ control::ControlStatus AudioRoutingController::connect_input_channel_to_track(in
     {
         auto status = _engine->connect_audio_input_channel(input_channel, track_channel, track_id);
         ELKLOG_LOG_ERROR_IF(status != EngineReturnStatus::OK, "Connecting audio channel {} to channel {} of track id {} failed with error {}",
-                input_channel, track_channel, track_id, static_cast<int>(status))
+                input_channel, track_channel, track_id, status)
 
         return status == EngineReturnStatus::OK? EventStatus::HANDLED_OK : EventStatus::ERROR;
     };
@@ -114,7 +114,7 @@ control::ControlStatus AudioRoutingController::connect_output_channel_to_track(i
     {
         auto status = _engine->connect_audio_output_channel(output_channel, track_channel, track_id);
         ELKLOG_LOG_ERROR_IF(status != EngineReturnStatus::OK, "Connecting audio channel {} from channel {} of track id {} failed with error {}",
-                           output_channel, track_channel, track_id, static_cast<int>(status))
+                           output_channel, track_channel, track_id, status)
 
         return status == EngineReturnStatus::OK? EventStatus::HANDLED_OK : EventStatus::ERROR;
     };
@@ -131,7 +131,7 @@ control::ControlStatus AudioRoutingController::disconnect_input(int track_id, in
     {
         auto status = _engine->disconnect_audio_input_channel(input_channel, track_channel, track_id);
         ELKLOG_LOG_ERROR_IF(status != EngineReturnStatus::OK, "Disconnecting audio channel {} to channel {} of track id {} failed with error {}",
-                           input_channel, track_channel, track_id, static_cast<int>(status))
+                           input_channel, track_channel, track_id, status)
 
         return status == EngineReturnStatus::OK? EventStatus::HANDLED_OK : EventStatus::ERROR;
     };
@@ -148,7 +148,7 @@ control::ControlStatus AudioRoutingController::disconnect_output(int track_id, i
     {
         auto status = _engine->disconnect_audio_output_channel(output_channel, track_channel, track_id);
         ELKLOG_LOG_ERROR_IF(status != EngineReturnStatus::OK, "Disconnecting audio channel {} from channel {} of track id {} failed with error {}",
-                           output_channel, track_channel, track_id, static_cast<int>(status))
+                           output_channel, track_channel, track_id, status)
 
         return status == EngineReturnStatus::OK? EventStatus::HANDLED_OK : EventStatus::ERROR;
     };
@@ -173,7 +173,7 @@ control::ControlStatus AudioRoutingController::disconnect_all_inputs_from_track(
                                                                       connection.track_channel,
                                                                       connection.track);
                 ELKLOG_LOG_ERROR_IF(status != EngineReturnStatus::OK, "Disconnecting audio channel {} from channel {} of track id {} failed with error {}",
-                                   connection.engine_channel, connection.track_channel, connection.track, static_cast<int>(status))
+                                   connection.engine_channel, connection.track_channel, connection.track, status)
 
                 return_status = status == EngineReturnStatus::OK ? return_status : EventStatus::ERROR;
             }
@@ -201,7 +201,7 @@ control::ControlStatus AudioRoutingController::disconnect_all_outputs_from_track
                                                                        connection.track_channel,
                                                                        connection.track);
                 ELKLOG_LOG_ERROR_IF(status != EngineReturnStatus::OK, "Disconnecting audio channel {} from channel {} of track id {} failed with error {}",
-                                   connection.engine_channel, connection.track_channel, connection.track, static_cast<int>(status))
+                                   connection.engine_channel, connection.track_channel, connection.track, status)
                 return_status = status == EngineReturnStatus::OK ? return_status : EventStatus::ERROR;
             }
         }
